requirements.txt
GitPython==3.1.31
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.1.1
